Mainly worked on calibration. Some ui stuff as well. Removed 22k Resistor.

+++ Punktschweissgeraet/Punktschweissgeraet/MeasureVoltage.h
@@ -20,7 +20,7 @@
 
 void initADC ();
 
-uint16_t readCapVoltage(volatile uint8_t number);
+float readCapVoltage(volatile uint8_t number);
 
 
 #endif /* MEASUREVOLTAGE_H_ */
